Arrow and Sword Damage based on stuff
Now based on stats. Formulla needs balancing but framework is there. Now
just need to do stuff with dexterity (and maybe add defense).

+++ Dungeon^3/Assets/Scripts/Enemies/Enemy.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Enemy : MonoBehaviour {
 
-	public int health;
+	public float health;
 	public int damage;
 	public Rigidbody rBody;
 	public GameObject player;
@@ -26,7 +26,7 @@ public class Enemy : MonoBehaviour {
 	
 	}
 
-	public void Damage(int number) {
+	public void Damage(float number) {
 		health -= number;
 		//Debug.Log(health);
 		if (health <= 0) {
